Initial commit. Added Dan's paint application.
I inted to modify the paint app to draw with lines instead of points. Then I will add accelerometer drawing, and then classes to deal specifically with sockets.

All of this will be organized into 3 branches, each branch for three versions of the app. This way, students can check out the branches for all to see!