Cleaning up all the code with comments. Simplified the drawing code when it comes to variable clearing.
I'm going to break down the drawing code some more, and really simplify the logic.

It's hard to follow currently, and I use too many variables. There should be a state variable that I manipulate for different drawing types. I'm also going to drop the toroidal aspect for the sake of readability.